Reasoning behind crossing off TP2 Task#1
Cross-OS compatibility seems to be an issue. ANSI escape characters won't work on Windows. Using terminal text-based GUIs such as curses would fix the issue, except it seems we have too many things entangled at this point if we add it. Quickly tried it and figured it was gonna take a while to get it to work in the project so I just thought about dropping it entirely.